+++ app/server/appsmith-server/src/main/java/com/appsmith/server/solutions/ce/ContextPermissionCE.java
@@ -1,0 +1,14 @@
+package com.appsmith.server.solutions.ce;
+
+import com.appsmith.server.acl.AclPermission;
+
+public interface ContextPermissionCE {
+
+    AclPermission getDeletePermission();
+
+    AclPermission getEditPermission();
+
+    default AclPermission getActionCreatePermission() {
+        return null;
+    }
+}
