+++ app/server/appsmith-server/src/main/java/com/appsmith/server/solutions/ce/PagePermissionCE.java
@@ -1,9 +1,5 @@
 package com.appsmith.server.solutions.ce;
 
-import com.appsmith.server.acl.AclPermission;
+import com.appsmith.server.solutions.ContextPermission;
 
-public interface PagePermissionCE {
-    AclPermission getDeletePermission();
-
-    AclPermission getActionCreatePermission();
-}
+public interface PagePermissionCE extends ContextPermission {}
