+++ app/server/appsmith-server/src/test/java/com/appsmith/server/helpers/ce/ImportArtifactPermissionProviderTest.java
@@ -9,6 +9,7 @@ import com.appsmith.server.domains.Application;
 import com.appsmith.server.domains.NewAction;
 import com.appsmith.server.domains.NewPage;
 import com.appsmith.server.domains.Workspace;
+import com.appsmith.server.helpers.ImportArtifactPermissionProvider;
 import com.appsmith.server.solutions.ActionPermission;
 import com.appsmith.server.solutions.ApplicationPermission;
 import com.appsmith.server.solutions.DatasourcePermission;
@@ -140,7 +141,7 @@ class ImportArtifactPermissionProviderTest {
         assertTrue(builder.permissionRequiredToCreateAction(true).build().isPermissionRequiredToCreateAction());
 
         assertTrue(builder.permissionRequiredToEditDatasource(true).build().isPermissionRequiredToEditDatasource());
-        assertTrue(builder.permissionRequiredToEditPage(true).build().isPermissionRequiredToEditPage());
+        assertTrue(builder.permissionRequiredToEditContext(true).build().isPermissionRequiredToEditContext());
         assertTrue(builder.permissionRequiredToEditAction(true).build().isPermissionRequiredToEditAction());
     }
 
@@ -159,7 +160,7 @@ class ImportArtifactPermissionProviderTest {
         assertTrue(provider.isPermissionRequiredToCreatePage());
         assertTrue(provider.isPermissionRequiredToCreateAction());
         assertTrue(provider.isPermissionRequiredToEditDatasource());
-        assertTrue(provider.isPermissionRequiredToEditPage());
+        assertTrue(provider.isPermissionRequiredToEditContext());
         assertTrue(provider.isPermissionRequiredToEditAction());
     }
 
@@ -188,7 +189,7 @@ class ImportArtifactPermissionProviderTest {
                 .currentUserPermissionGroups(Set.of());
 
         if (baseDomain instanceof NewPage) {
-            builder.permissionRequiredToEditPage(true);
+            builder.permissionRequiredToEditContext(true);
         } else if (baseDomain instanceof NewAction) {
             builder.permissionRequiredToEditAction(true);
         } else if (baseDomain instanceof Datasource) {
